fix(tests): avoid maintenance audit self-starvation
