+++ src/tavily_proxy/proxy_quota_sync_and_jobs.rs
@@ -3039,40 +3039,3 @@ fn normalize_quota_sync_fetch_error(err: ProxyError) -> ProxyError {
         other => other,
     }
 }
-
-#[cfg(test)]
-mod privacy_status_phase_budget_tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn privacy_status_stops_at_a_safe_boundary_and_closes_its_snapshot() {
-        let temp_dir = tempfile::tempdir().expect("temp dir");
-        let db_path = temp_dir.path().join("privacy-status-phase-budget.db");
-        let db_str = db_path.to_string_lossy().to_string();
-        let proxy = TavilyProxy::with_endpoint(
-            vec!["tvly-privacy-status-phase-budget".to_string()],
-            crate::DEFAULT_UPSTREAM,
-            &db_str,
-        )
-        .await
-        .expect("proxy created");
-
-        let error = proxy
-            .upstream_privacy_status_after_one_safe_boundary_for_test()
-            .await
-            .expect_err("the real status builder must stop before its second SQLite phase");
-        assert!(matches!(
-            error,
-            ProxyError::Database(sqlx::Error::PoolTimedOut)
-        ));
-        assert_eq!(
-            proxy.admin_privacy_read_discards_for_test(),
-            0,
-            "phase-bound refresh aborts must close the snapshot without discarding it"
-        );
-        proxy
-            .verify_admin_privacy_read_connection_clean_for_test()
-            .await
-            .expect("the next privacy transaction is clean after a phase-bound refresh abort");
-    }
-}
